third_party/WebKit/LayoutTests/editing/pasteboard/smart-paste-007.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div class="explanation">
<div class="scenario">
Tests: 
<br>
Smart paste when pasting between two periods.
</div>
<div class="expected-results">
Expected Results:
<br>
A space should be added before, but not after the pasted word. It should look like this:
<BR>
. test.
</div>
</div>

<div contenteditable id="root" style="word-wrap: break-word; -khtml-nbsp-mode: space; -khtml-line-break: after-white-space;"><div id="sample">test</div></div>
<div id="log"></div>
<script>
test(function() {
    if (!window.internals && !window.eventSender)
        return;

    internals.settings.setEditingBehavior('win');

    var selection = window.getSelection();
    var sample = document.getElementById('sample');

    var x = sample.offsetLeft;
    var y = sample.offsetTop + sample.offsetHeight / 2;
    eventSender.mouseMoveTo(x, y);
    eventSender.mouseDown();
    eventSender.mouseUp();
    eventSender.mouseDown();
    eventSender.mouseUp();

    document.execCommand('cut');
    document.execCommand('insertText', false, '.');
    document.execCommand('insertText', false, '.');
    selection.modify('move', 'backward', 'character');
    document.execCommand('paste');

    assert_equals(sample.innerHTML, '. test.');
    assert_true(selection.isCollapsed, 'isCollapsed');
    assert_equals(selection.anchorNode, sample.firstChild, 'anchorNode');
    assert_equals(selection.anchorOffset, 6, 'anchorOffset');
});
</script>
